diet_planner_agent/requirements.txt
google-adk==1.18.0
pytest==8.4.2
pytest-asyncio==1.2.0
pandas
tabulate
tqdm
scikit-learn
deprecated
